+++ cmd/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/keptn-contrib/dynatrace-service/internal/credentials"
 	"github.com/keptn-contrib/dynatrace-service/internal/env"
 	"github.com/keptn-contrib/dynatrace-service/internal/event_handler"
-	"github.com/keptn-contrib/dynatrace-service/internal/http_handler"
+	"github.com/keptn-contrib/dynatrace-service/internal/health"
 	"github.com/keptn-contrib/dynatrace-service/internal/onboard"
 
 	log "github.com/sirupsen/logrus"
@@ -47,7 +47,7 @@ func _main(args []string, envCfg envConfig) int {
 	ctx = cloudevents.WithEncodingStructured(ctx)
 
 	log.WithFields(log.Fields{"port": envCfg.Port, "path": envCfg.Path}).Debug("Initializing cloudevents client")
-	p, err := cloudevents.NewHTTP(cloudevents.WithPath(envCfg.Path), cloudevents.WithPort(envCfg.Port), cloudevents.WithGetHandlerFunc(http_handler.HTTPGetHandler))
+	p, err := cloudevents.NewHTTP(cloudevents.WithPath(envCfg.Path), cloudevents.WithPort(envCfg.Port), cloudevents.WithGetHandlerFunc(health.HTTPGetHandler))
 	if err != nil {
 		log.WithError(err).Fatal("Failed to create client")
 	}
