+++ internal/health/http_get_handler.go
@@ -1,4 +1,4 @@
-package http_handler
+package health
 
 import (
 	"net/http"
